Update excludes for new Frechet tests 102 (#103)
* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...


Former-commit-id: 8f32a6be6ef3cf4cd8b9dc3af1e8a338926eb0ca
Former-commit-id: dd576a5c67230a9d65b2e1128c570401e3bc4a4e
Former-commit-id: 078904c847f825a15deee779c1acc2a92b964d1d
Former-commit-id: 0ffe8e5b7dd76f60e1a2afe52007df7db8b29133
Former-commit-id: 46c2b2822bc92b17446e4b2a776573b4bfa07a9b