Update excludes for new Frechet tests 102 (#103)
* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...


Former-commit-id: 369c1b2ec3966b7f24054cfa3716ca5781499f2e
Former-commit-id: 63b8b6d5f0734203755bc5743f91563b4205c000
Former-commit-id: 74570e85d1a7ad0a5cf72a48cb6432b4c90fe0f2
Former-commit-id: cf18492dc288d9df85020d6949d800e7449f25a2
Former-commit-id: 2c4a469a709303c119466673331be8d6e5b428ef