Update excludes for new Frechet tests 102 (#103)
* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...


Former-commit-id: f78a6e9c6747d17b0c05494c3969aca508c8ecd1
Former-commit-id: 6e38dc1709cce2abe96a3852dcb10228c7330620
Former-commit-id: 67fb33025dab5d0ab0923802d706b4051509d02c
Former-commit-id: 06b816798f4d645f53732829fe27df0f4b13a23a
Former-commit-id: a88618bf413f167472aaa9525740f76be0e2502b