Merge latest jdk 131 fix into develop (#149)
* Update excludes for new Frechet tests 102 (#103)

* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...

* Fix jdk requires to latest release


Former-commit-id: 678d6fd30004a0cf4f304b26dc57c9a50ae5c5c6
Former-commit-id: a2962d80279c0d184c21c40738f26bb365b3eb69
Former-commit-id: fd88a08532964b481bb4d9f3006b6f5941aeaf0f
Former-commit-id: 8800c25aa2b1d88fc9275c875d7d30c039b1832a
Former-commit-id: 788c8b3147c6e721a1a05cdcb7ab828526047e29